requirements.txt
######################################################################################################################################################

# @project        Space Elevator ▸ LEDs Strip
# @file           requirements.txt
# @author         <info@spacelevator.org>
# @license        Space Elevator 2022

######################################################################################################################################################

pyserial==3.5
requests==2.25.1

######################################################################################################################################################
